+++ lib/extensionScripts/help.d.ts
@@ -1,0 +1,2 @@
+declare const _default: "\nzeroconfig-extension-scripts watch app\nzeroconfig-extension-scripts build app\n\n--static-file-directories \"public static ...\"\n--static-file-packages \"package1 package2 ...\"\n--output /path/to\n--vendor-file-name vendor\n--style-file-name style";
+export default _default;