+++ lib/types.d.ts
@@ -2,14 +2,17 @@ import { PackageJson } from 'type-fest';
 export declare const packageCommands: readonly ["build", "publish", "validate", "list", "sync"];
 export declare const webappCommands: readonly ["build", "start", "server-watch", "server-start", "browser-start"];
 export declare const desktopappCommands: readonly ["build", "start", "electron-watch", "electron-start"];
+export declare const extensionCommands: readonly ["build", "watch"];
 export declare const modes: readonly ["production", "development"];
 export declare type PackageCommand = typeof packageCommands[number];
 export declare type WebappCommand = typeof webappCommands[number];
 export declare type DesktopappCommand = typeof desktopappCommands[number];
+export declare type ExtensionCommand = typeof extensionCommands[number];
 export declare type Mode = typeof modes[number];
 export declare function isPackageCommand(command: string): command is PackageCommand;
 export declare function isWebappCommand(command: string): command is WebappCommand;
 export declare function isDesktopappCommand(command: string): command is DesktopappCommand;
+export declare function isExtensionCommand(command: string): command is ExtensionCommand;
 export declare function isMode(mode: string | undefined): mode is Mode;
 export interface PackageArgv {
     command: PackageCommand;
@@ -96,6 +99,29 @@ export interface DesktopappConfig {
     app: string;
     staticFileDirectories: string[];
     output: string;
+    cwd: string;
+    zeroconfigPath: string;
+    extend: {
+        templateFiles: string[];
+    };
+}
+export interface ExtensionArgv {
+    command: ExtensionCommand;
+    app: string;
+    output: string | undefined;
+    vendorFileName: string;
+    styleFileName: string;
+    staticFileDirectories: string | undefined;
+    staticFilePackages: string | undefined;
+}
+export interface ExtensionConfig {
+    command: ExtensionCommand;
+    app: string;
+    output: string;
+    vendorFileName: string;
+    styleFileName: string;
+    staticFileDirectories: string[];
+    entryFiles: string[];
     cwd: string;
     zeroconfigPath: string;
     extend: {