+++ src/utils/sayZeroconfig.ts
@@ -4,7 +4,7 @@ import path from 'path';
 import { PackageJson } from 'type-fest';
 
 export function sayZeroconfig() {
-  say('ZEROCONFIG', {font: 'block'});
+  say('ZERO\nCONFIG', {font: 'block'});
   
   const {version}: PackageJson = fs.readJsonSync(path.join(__dirname, '../../package.json'));